Fix compile error in GHC 9.2
The ConP constructor takes an extra argument in GHC-9.2, and this fix ensures that Heterocephalus compiles for GHC-9.2 whilst continuing to be backwards compatible with pre GHC-9.2. Note that this patch does not introduce logic to handle the new GHC-9.2 syntax (whatever that is), it just ensures that the current behaviour pre GHC-9.2 features can continue to compile under GHC-9.2.